bench: Generalize wave3d
